refactor(trip-inspection): relocate parts to trip dir
Move the trip-inspection auxiliary components into src/components/trip/ and drop the TripInspection prefix so file names line up with the new location:

  trip-basic-info.tsx (moved)
  trip-inspection-trip-pager.tsx -> trip-pager.tsx
    TripInspectionTripPager -> TripPager
  trip-inspection-stop-list.tsx -> trip-stops.tsx
    TripInspectionStopList -> TripStops
    TripInspectionStopRow -> TripStopRow
    TripInspectionPlaceholderRow -> TripStopPlaceholderRow
    TripInspectionMetaInfo -> TripStopMetaInfo

Swap StopTimeDetailInfo for TripInfo inside TripStopRow so each row renders the trip-centric info layout.
