+++ src/components/trip/trip-basic-info.tsx
